+++ src/main/java/com/jonescorp/shoppy/repository/TransactionRepository.java
@@ -3,5 +3,5 @@ package com.jonescorp.shoppy.repository;
 import com.jonescorp.shoppy.model.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {
 }
